Refactor: [M3-6280]: Two factor
