+++ packages/manager/src/features/Profile/AuthenticationSettings/TwoFactor/EnableTwoFactorForm.tsx
@@ -1,27 +1,25 @@
-import { confirmTwoFactor } from '@linode/api-v4/lib/profile';
-import { APIError } from '@linode/api-v4/lib/types';
 import * as React from 'react';
-import { CircleProgress } from 'src/components/CircleProgress';
-import Divider from 'src/components/core/Divider';
-import { Notice } from 'src/components/Notice/Notice';
-import { getAPIErrorOrDefault, getErrorMap } from 'src/utilities/errorUtils';
-import scrollErrorIntoView from 'src/utilities/scrollErrorIntoView';
 import ConfirmToken from './ConfirmToken';
+import Divider from 'src/components/core/Divider';
+import scrollErrorIntoView from 'src/utilities/scrollErrorIntoView';
+import { APIError } from '@linode/api-v4/lib/types';
+import { CircleProgress } from 'src/components/CircleProgress';
+import { confirmTwoFactor } from '@linode/api-v4/lib/profile';
+import { getAPIErrorOrDefault, getErrorMap } from 'src/utilities/errorUtils';
+import { Notice } from 'src/components/Notice/Notice';
 import { QRCodeForm } from './QRCodeForm';
 
 interface Props {
   loading: boolean;
-  secret: string;
-  username: string;
-  twoFactorConfirmed: boolean;
-  onSuccess: (scratchCode: string) => void;
   onCancel: () => void;
+  onSuccess: (scratchCode: string) => void;
+  secret: string;
   toggleDialog: () => void;
+  twoFactorConfirmed: boolean;
+  username: string;
 }
 
-type CombinedProps = Props;
-
-export const EnableTwoFactorForm: React.FC<CombinedProps> = (props) => {
+export const EnableTwoFactorForm = (props: Props) => {
   const [errors, setErrors] = React.useState<APIError[] | undefined>(undefined);
   const [submitting, setSubmitting] = React.useState<boolean>(false);
   const [token, setToken] = React.useState<string>('');
@@ -98,5 +96,3 @@ export const EnableTwoFactorForm: React.FC<CombinedProps> = (props) => {
     </React.Fragment>
   );
 };
-
-export default EnableTwoFactorForm;
