+++ packages/manager/src/features/Profile/AuthenticationSettings/TwoFactor/TwoFactor.tsx
@@ -1,46 +1,19 @@
-import { getTFAToken } from '@linode/api-v4/lib/profile';
-import { APIError } from '@linode/api-v4/lib/types';
 import * as React from 'react';
-import FormControl from 'src/components/core/FormControl';
-import FormControlLabel from 'src/components/core/FormControlLabel';
-import { makeStyles } from 'tss-react/mui';
-import { Theme } from '@mui/material/styles';
-import Typography from 'src/components/core/Typography';
-import { Notice } from 'src/components/Notice/Notice';
-import { Toggle } from 'src/components/Toggle';
-import { queryKey } from 'src/queries/profile';
-import { useSecurityQuestions } from 'src/queries/securityQuestions';
-import { getAPIErrorOrDefault } from 'src/utilities/errorUtils';
-import getAPIErrorFor from 'src/utilities/getAPIErrorFor';
 import DisableTwoFactorDialog from './DisableTwoFactorDialog';
-import EnableTwoFactorForm from './EnableTwoFactorForm';
+import getAPIErrorFor from 'src/utilities/getAPIErrorFor';
 import ScratchDialog from './ScratchCodeDialog';
+import Typography from 'src/components/core/Typography';
+import { APIError } from '@linode/api-v4/lib/types';
+import { EnableTwoFactorForm } from './EnableTwoFactorForm';
+import { getAPIErrorOrDefault } from 'src/utilities/errorUtils';
+import { getTFAToken } from '@linode/api-v4/lib/profile';
+import { Notice } from 'src/components/Notice/Notice';
+import { queryKey } from 'src/queries/profile';
+import { LinkStyledButton } from 'src/components/Button/LinkStyledButton';
+import { TwoFactorToggle } from './TwoFactorToggle';
 import { useQueryClient } from 'react-query';
-
-const useStyles = makeStyles()((theme: Theme) => ({
-  container: {
-    display: 'flex',
-    flexFlow: 'row nowrap',
-    alignItems: 'center',
-    justifyContent: 'left',
-    marginTop: theme.spacing(3),
-    marginBottom: theme.spacing(3),
-  },
-  copy: {
-    lineHeight: '20px',
-    marginTop: theme.spacing(),
-    marginBottom: theme.spacing(),
-    maxWidth: 960,
-  },
-  button: {
-    ...theme.applyLinkStyles,
-  },
-  disabled: {
-    '& *': {
-      color: theme.color.disabledText,
-    },
-  },
-}));
+import { useSecurityQuestions } from 'src/queries/securityQuestions';
+import { StyledRootContainer, StyledCopy, useStyles } from './TwoFactor.styles';
 
 interface Props {
   username?: string;
@@ -50,14 +23,10 @@ interface Props {
 
 export const TwoFactor: React.FC<Props> = (props) => {
   const { classes } = useStyles();
-
   const needSecurityQuestionsCopy =
     'To use two-factor authentication you must set up your security questions listed below.';
-
   const { disabled, twoFactor, username } = props;
-
   const queryClient = useQueryClient();
-
   const [errors, setErrors] = React.useState<APIError[] | undefined>(undefined);
   const [loading, setLoading] = React.useState<boolean>(false);
   const [secret, setSecret] = React.useState<string>('');
@@ -205,12 +174,12 @@ export const TwoFactor: React.FC<Props> = (props) => {
             text={generalError}
           />
         )}
-        <Typography variant="body1" className={classes.copy} data-qa-copy>
+        <StyledCopy variant="body1" data-qa-copy>
           Two-factor authentication increases the security of your Cloud Manager
           account by requiring two different forms of authentication to log in:
           your Cloud Manager account password and an authorized security token
           generated by another platform.
-        </Typography>
+        </StyledCopy>
         {(hasSecurityQuestions && !twoFactor) || twoFactor ? (
           typeof twoFactorConfirmed !== 'undefined' && (
             <TwoFactorToggle
@@ -230,19 +199,15 @@ export const TwoFactor: React.FC<Props> = (props) => {
           />
         )}
         {twoFactorEnabled && (
-          <div className={classes.container}>
-            <button
-              className={classes.button}
-              onClick={toggleHidden}
-              data-qa-hide-show-code
-            >
+          <StyledRootContainer>
+            <LinkStyledButton onClick={toggleHidden} data-qa-hide-show-code>
               {showQRCode
                 ? 'Hide QR Code'
                 : twoFactorConfirmed
                 ? 'Reset two-factor authentication'
                 : 'Show QR Code'}
-            </button>
-          </div>
+            </LinkStyledButton>
+          </StyledRootContainer>
         )}
         {twoFactorEnabled &&
           showQRCode &&
@@ -270,49 +235,5 @@ export const TwoFactor: React.FC<Props> = (props) => {
         closeDialog={toggleDisable2FA}
       />
     </React.Fragment>
-  );
-};
-
-export default TwoFactor;
-
-interface ToggleProps {
-  toggleDisableDialog: () => void;
-  onChange: (value: boolean) => void;
-  twoFactorEnabled: boolean;
-  twoFactorConfirmed: boolean;
-  disabled?: boolean;
-}
-
-const TwoFactorToggle: React.FC<ToggleProps> = (props) => {
-  const { disabled, twoFactorEnabled } = props;
-
-  const handleChange = (e: React.ChangeEvent<HTMLInputElement>) => {
-    const { twoFactorConfirmed, onChange } = props;
-    const enabled = e.currentTarget.checked;
-    /**
-     * only open the disable dialog if 2FA has been turned on and we're flipping the toggle off
-     */
-    if (!enabled && twoFactorConfirmed) {
-      props.toggleDisableDialog();
-    } else {
-      /** Otherwise flip the toggle. If toggling on, the parent will handle the API request. */
-      onChange(enabled);
-    }
-  };
-
-  return (
-    <FormControl fullWidth style={{ marginTop: 0 }}>
-      <FormControlLabel
-        label={twoFactorEnabled ? 'Enabled' : 'Disabled'}
-        control={
-          <Toggle
-            checked={twoFactorEnabled}
-            onChange={handleChange}
-            data-qa-toggle-tfa={twoFactorEnabled}
-            disabled={disabled}
-          />
-        }
-      />
-    </FormControl>
   );
 };
