--- packages/manager/src/features/Profile/AuthenticationSettings/TwoFactor/index.tsx
@@ -1,2 +1,0 @@
-import TwoFactor from './TwoFactor';
-export default TwoFactor;
